Simplify chart wrapper and plug in simple-chart
